Support for large mailboxes, now with `long long`.

+++ mairix.h
@@ -327,7 +327,7 @@ enum data_to_rfc822_error {
   DTR8_BAD_HEADERS, /* corrupt headers */
   DTR8_BAD_ATTACHMENT /* corrupt attachment (e.g. no body part) */
 };
-struct rfc822 *data_to_rfc822(struct msg_src *src, char *data, long length, enum data_to_rfc822_error *error);
+struct rfc822 *data_to_rfc822(struct msg_src *src, char *data, size_t length, enum data_to_rfc822_error *error);
 void create_ro_mapping(const char *filename, unsigned char **data, size_t *len);
 void free_ro_mapping(unsigned char *data, size_t len);
 char *format_msg_src(struct msg_src *src);
